Consolidate diff.js and diff-match-patch.js
Use the newer version of the library from diff-match-patch.js, but
remove match/patch functions since we don't need them either.
